Refactor: separate edit colony action saga
