+++ src/modules/dashboard/sagas/actions/index.ts
@@ -6,6 +6,7 @@ import mintTokensActionSaga from './mintTokens';
 import versionUpgradeActionSaga from './versionUpgrade';
 import createDomainActionSaga from './createDomain';
 import editDomainActionSaga from './editDomain';
+import editColonyActionSaga from './editColony';
 
 export default function* actionsSagas() {
   yield all([
@@ -15,5 +16,6 @@ export default function* actionsSagas() {
     call(versionUpgradeActionSaga),
     call(createDomainActionSaga),
     call(editDomainActionSaga),
+    call(editColonyActionSaga),
   ]);
 }
